Move compare and buildDiff back to main process
