+++ lib/image.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var q = require('q'),
+    looksSame = require('looks-same'),
     PngImg = require('png-img'),
     utils = require('png-img/utils'),
     inherit = require('inherit');
@@ -43,7 +44,31 @@ var Image = inherit({
 
     RGBToString: function(rgb) {
         return utils.RGBToString(rgb);
-    }
+    },
+
+    compare: function(path1, path2, opts) {
+        opts = opts || {};
+        var compareOptions = {
+            ignoreCaret: opts.canHaveCaret
+        };
+        if ('tolerance' in opts) {
+            compareOptions.tolerance = opts.tolerance;
+        }
+        return q.nfcall(looksSame, path1, path2, compareOptions);
+    },
+
+    buildDiff: function(opts) {
+        var diffOptions = {
+            reference: opts.reference,
+            current: opts.current,
+            diff: opts.diff,
+            highlightColor: opts.diffColor
+        };
+        if ('tolerance' in opts) {
+            diffOptions.tolerance = opts.tolerance;
+        }
+        return q.nfcall(looksSame.createDiff, diffOptions);
+    },
 });
 
 module.exports = Image;
