+++ lib/state-processor/capture-processor/capture-processor.js
@@ -1,15 +1,10 @@
 'use strict';
 
-var inherit = require('inherit'),
-    q = require('q');
+var inherit = require('inherit');
 
 module.exports = inherit({
     getProcessedCaptureEventName: function() {
         throw new Error('Not implemented');
-    },
-
-    prepare: function(emitter) {
-        return q();
     },
 
     exec: function(capture, opts) {
