+++ lib/state-processor/capture-processor/screen-updater/diff-screen-updater.js
@@ -4,19 +4,10 @@ var fs = require('q-io/fs'),
     inherit = require('inherit'),
     temp = require('../../../temp'),
 
-    ImageProcessor = require('../../../image-processor'),
+    Image = require('../../../image'),
     ScreenUpdater = require('./screen-updater');
 
 module.exports = inherit(ScreenUpdater, {
-    __constructor: function() {
-        this._imageProcessor = null;
-    },
-
-    prepare: function(emitter) {
-        this._imageProcessor = new ImageProcessor(emitter);
-        return this.__base(emitter);
-    },
-
     _processCapture: function(capture, opts, isRefExists) {
         if (!isRefExists) {
             return;
@@ -28,7 +19,7 @@ module.exports = inherit(ScreenUpdater, {
 
         return capture.image.save(tmpPath)
             .then(function() {
-                return this._imageProcessor.compare(tmpPath, refPath, {
+                return Image.compare(tmpPath, refPath, {
                         canHaveCaret: capture.canHaveCaret,
                         tolerance: tolerance
                     })
