+++ lib/state-processor/capture-processor/screen-updater/meta-screen-updater.js
@@ -1,7 +1,6 @@
 'use strict';
 
-var q = require('q'),
-    inherit = require('inherit'),
+var inherit = require('inherit'),
     ScreenUpdater = require('./screen-updater'),
     DiffUpdater = require('./diff-screen-updater'),
     NewUpdater = require('./new-screen-updater'),
@@ -12,13 +11,6 @@ module.exports = inherit(ScreenUpdater, {
     __constructor: function() {
         this._diffUpdater = new DiffUpdater();
         this._newUpdater = new NewUpdater();
-    },
-
-    prepare: function(emitter) {
-        return q.all([
-            this._diffUpdater.prepare(emitter),
-            this._newUpdater.prepare(emitter)
-        ]);
     },
 
     _processCapture: function(capture, opts, isRefExists) {
