+++ lib/state-processor/capture-processor/tester.js
@@ -6,7 +6,7 @@ var q = require('q'),
     temp = require('../../temp'),
 
     CaptureProcessor = require('./capture-processor'),
-    ImageProcessor = require('../../image-processor'),
+    Image = require('../../image'),
     NoRefImageError = require('../../errors/no-ref-image-error'),
 
     RunnerEvents = require('../../constants/runner-events');
@@ -14,16 +14,10 @@ var q = require('q'),
 module.exports = inherit(CaptureProcessor, {
     __constructor: function(diffColor) {
         this._diffColor = diffColor;
-        this._imageProcessor = null;
     },
 
     getProcessedCaptureEventName: function() {
         return RunnerEvents.END_TEST;
-    },
-
-    prepare: function(emitter) {
-        this._imageProcessor = new ImageProcessor(emitter);
-        return this.__base(emitter);
     },
 
     exec: function(capture, opts) {
@@ -42,7 +36,7 @@ module.exports = inherit(CaptureProcessor, {
                 }
             })
             .then(function() {
-                return [tmpPath, _this._imageProcessor.compare(tmpPath, refPath, {
+                return [tmpPath, Image.compare(tmpPath, refPath, {
                     canHaveCaret: capture.canHaveCaret,
                     tolerance: tolerance
                 })];
@@ -53,7 +47,7 @@ module.exports = inherit(CaptureProcessor, {
                     currentPath: currentPath,
                     equal: isEqual,
                     saveDiffTo: function(diffPath) {
-                        return _this._imageProcessor.buildDiff({
+                        return Image.buildDiff({
                             reference: refPath,
                             current: currentPath,
                             diff: diffPath,
