+++ lib/state-processor/state-processor.js
@@ -1,7 +1,8 @@
 'use strict';
 
 var inherit = require('inherit'),
-    _ = require('lodash');
+    _ = require('lodash'),
+    q = require('q');
 
 module.exports = inherit({
     __constructor: function(captureProcessor) {
@@ -13,7 +14,7 @@ module.exports = inherit({
     },
 
     prepare: function(emitter) {
-        return this._captureProcessor.prepare(emitter);
+        return q();
     },
 
     exec: function(state, browserSession, pageDisposition) {
