+++ test/functional/image.test.js
@@ -56,4 +56,67 @@ describe('image', function() {
             assert.deepEqual(this.image.getRGBA(0, 0), {r: 9, g: 9, b: 9, a: 255});
         });
     });
+
+    describe('compare', function() {
+        it('should resolve to `true` for equal images', function() {
+            return assert.eventually.isTrue(Image.compare(
+                util.imagePath('image1.png'),
+                util.imagePath('image2.png')
+            ));
+        });
+
+        it('should resolve to `false` for non-equal images', function() {
+            return assert.eventually.isFalse(Image.compare(
+                util.imagePath('image1.png'),
+                util.imagePath('image3.png')
+            ));
+        });
+
+        it('should resolve to `true` for non-equal images if tolerance is high enough', function() {
+            return assert.eventually.isTrue(Image.compare(
+                util.imagePath('image1.png'),
+                util.imagePath('image3.png'),
+                {tolerance: 50}
+            ));
+        });
+
+        it('should resolve to `true` for images with unnoticable difference', function() {
+            return assert.eventually.isTrue(Image.compare(
+                util.imagePath('image1.png'),
+                util.imagePath('image4.png')
+            ));
+        });
+    });
+
+    describe('buildDiff', function() {
+        it('should build diff image', function() {
+            return util.withTempFile(function(fileName) {
+                var opts = {
+                    reference: util.imagePath('image1.png'),
+                    current: util.imagePath('image3.png'),
+                    diff: fileName,
+                    diffColor: '#f0001c'
+                };
+                return Image.buildDiff(opts)
+                    .then(function() {
+                        return util.assertSameImages('image_diff.png', fileName);
+                    });
+            });
+        });
+
+        it('should allow to change diff color', function() {
+            return util.withTempFile(function(fileName) {
+                var opts = {
+                    reference: util.imagePath('image1.png'),
+                    current: util.imagePath('image3.png'),
+                    diff: fileName,
+                    diffColor: '#0000ff'
+                };
+                return Image.buildDiff(opts)
+                    .then(function() {
+                        return util.assertSameImages('image_diff_blue.png', fileName);
+                    });
+            });
+        });
+    });
 });
